rework data-connections
Add a new struct network-segment to handle the data-connections
for the nodes and synapses of all brick in one single buffer.
This should make it easier when syncing the data with the gpu.
At the moment it compiles at least at the momemnt, but the
functionality is totally broken again.