version 0.2.0
Squashed the last about 600 commits to make the base clean,
because a huge amount of these commits are only try-and-error stuff,
which are very dirty and not really usefully anymore. These old
commits still only exist as backup.

The current commit contains version 0.2.0 of the code with basic
functionallity, even it can nothing really useful at the moment
and there are also many open questions about the scalability and
problems with calibration of some parammeters, but at least it
is a proof of the concept version 2.