restructure input-output-handline
Move old structres for input and output handling out of the
KyoukoRoot, NetworkManager and Brick class into its own class.
Instead of handling the input and output in the edge-processing,
this handling is now an extra step into the pipelining. Also
made better mapping available, even if only one output node
is supported at the moment.